+++ src/extension.ts
@@ -2,6 +2,7 @@ import { subscribeCommands } from "./commands";
 import { subscribeEventListeners } from "./eventListeners";
 import { executeOnExtensionActive } from "./executeOnExtensionActive";
 import { executeOnExtensionDeactive } from "./executeOnExtensionDeactive";
+import { registerMarkdown } from "./modules/markdown";
 import { subscribeProviders } from "./providers";
 import { vscode } from "./shared";
 import { extensionName, init } from "./shared/init";
@@ -16,10 +17,14 @@ export async function activate(context: vscode.ExtensionContext) {
         subscribeEventListeners();
         subscribeProviders();
 
+        registerMarkdown();
+
         console.log(`${extensionName} is now active!`);
     } catch (e) {
         console.error(e);
-        vscode.window.showErrorMessage(`${e}`);
+        vscode.window.showErrorMessage(
+            `Error while initializing extension. ${e}`
+        );
     }
 }