+++ src/modules/markdown/diagnostics/index.ts
@@ -1,0 +1,7 @@
+import { registerNoInvalidLocalImageFilePath } from "./noInvalidLocalImageFilePath";
+import { registerNoLocalImageLink } from "./noLocalImageLink";
+
+export function registerDiagnostics() {
+    registerNoInvalidLocalImageFilePath();
+    registerNoLocalImageLink();
+}