+++ src/modules/markdown/diagnostics/noInvalidLocalImageFilePath/index.ts
@@ -1,0 +1,72 @@
+import path from "node:path";
+
+import { fs, vscode } from "@/shared";
+import { extensionCtx, extensionName } from "@/shared/init";
+
+import { kLocalImageLinkRegex } from "../../consts";
+
+let diagnosticCollection: vscode.DiagnosticCollection;
+
+export function registerNoInvalidLocalImageFilePath() {
+    diagnosticCollection = vscode.languages.createDiagnosticCollection(
+        "markdown-local-image"
+    );
+    extensionCtx.subscriptions.push(diagnosticCollection);
+
+    extensionCtx.subscriptions.push(
+        vscode.workspace.onDidOpenTextDocument(appendDiagnostics)
+    );
+    extensionCtx.subscriptions.push(
+        vscode.workspace.onDidSaveTextDocument(appendDiagnostics)
+    );
+    extensionCtx.subscriptions.push(
+        vscode.workspace.onDidChangeTextDocument(e =>
+            appendDiagnostics(e.document)
+        )
+    );
+}
+
+function appendDiagnostics(document: vscode.TextDocument) {
+    if (document.languageId !== "markdown") {
+        return;
+    }
+
+    const diagnostics: vscode.Diagnostic[] = [];
+    const text = document.getText();
+
+    // Match all local image links and check if the image file exists
+
+    let match;
+    while ((match = kLocalImageLinkRegex.exec(text)) !== null) {
+        const [imageLink, imagePath] = match;
+
+        const imageAbsPath = path.resolve(
+            path.dirname(document.uri.fsPath),
+            imagePath
+        );
+        if (!fs.existsSync(imageAbsPath)) {
+            const imagePathStartPos = document.positionAt(
+                match.index + imageLink.indexOf(imagePath)
+            );
+            const imagePathEndPos = document.positionAt(
+                match.index + imageLink.indexOf(imagePath) + imagePath.length
+            );
+
+            const imagePathRange = new vscode.Range(
+                imagePathStartPos,
+                imagePathEndPos
+            );
+
+            const diagnostic = new vscode.Diagnostic(
+                imagePathRange,
+                `Local image file "${imagePath}" not found.`,
+                vscode.DiagnosticSeverity.Error
+            );
+            diagnostic.code = `@${extensionName}/no-invalid-local-image-file`;
+
+            diagnostics.push(diagnostic);
+        }
+    }
+
+    diagnosticCollection.set(document.uri, diagnostics);
+}