+++ src/modules/markdown/index.ts
@@ -1,0 +1,5 @@
+import { registerDiagnostics } from "./diagnostics";
+
+export function registerMarkdown() {
+    registerDiagnostics();
+}