+++ src/providers/codeActionProvider/commonCodeActionProvider.ts
@@ -10,8 +10,8 @@ export class CommonCodeActionProvider implements vscode.CodeActionProvider {
     provideCodeActions(
         document: vscode.TextDocument,
         range: vscode.Range | vscode.Selection,
-        _context: vscode.CodeActionContext,
-        _token: vscode.CancellationToken
+        context: vscode.CodeActionContext,
+        token: vscode.CancellationToken
     ): vscode.ProviderResult<(vscode.CodeAction | vscode.Command)[]> {
         /* timestampTool */