+++ CHANGELOG.md
@@ -5,11 +5,20 @@ All notable changes to "sower-tools" will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] - 2025-02-13
+
+### Added
+
+-   Add markdown image diagnostics, including:
+    -   `no-invalid-local-image-file-path`
+    -   `no-local-image-link`
+-   Add upload local image to minio storage functionality.
+
 ## [0.16.3] - 2025-01-24
 
 ### Changed
 
--   Update countdown timer functionality and update build tool using pnpm.
+-   Update countdown timer functionality.
 
 ## [0.16.2] - 2024-12-26